pyproject.toml
[tool.poetry]
name = "pythonproject2"
version = "0.1.0"
description = ""
authors = ["Oksana <oksanakysylychak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.16.1"
mysqlclient = "^2.2.7"
dotenv = "^0.9.9"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^11.3.0"



[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
